requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-3-jet==1.0.8
django-appconf==1.0.4
django-braces==1.14.0
django-ckeditor==6.0.0
django-clearcache==1.2.1
django-compressor==2.4
django-elasticsearch-dsl==7.2.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-prettyjson==0.4.1
django-rest-elasticsearch==0.4.2
django-utils-six==2.0
django-workers==0.2.0
djangorestframework==3.12.4
elasticsearch==7.12.1
elasticsearch-dsl==7.3.0
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
idna==2.10
joblib==1.0.1
Markdown==3.3.4
nltk==3.5
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.7
Pillow==8.1.2
psycopg2==2.8.6
Pygments==2.9.0
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-Levenshtein-wheels==0.13.2
pytz==2021.1
PyYAML==5.4.1
rarfile==4.0
rcssmin==1.0.6
regex==2021.3.17
requests==2.25.1
rjsmin==1.1.0
singlemodeladmin==0.9
six==1.15.0
sqlparse==0.4.1
standardjson==0.3.1
textblob==0.15.3
tqdm==4.59.0
Unidecode==1.2.0
urllib3==1.26.4
